+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.4.0] - 2017-09-26
+### Added
+- setLevel function
+- checking log level in main log function
+
 ## [0.3.0] - 2017-09-25
 ### Added
 - functionality for mixing of user option with user option